# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Disaster Map API
  description: |
    Real-time 3D mapping and visualization API for disaster response operations.
    
    This API provides endpoints for:
    - Video stream ingestion and frame processing
    - SLAM pose estimation and tracking
    - Point cloud map management
    - Live data streaming via WebSocket
    
    **Base URL:** `https://api.disaster-map.local/v1`
  version: 1.0.0
  contact:
    name: Disaster Map Team
    email: support@disaster-map.local
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.disaster-map.local/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Local development server
...
paths:
...
  /poses:
    get:
      summary: Get pose estimates
      description: |
        Retrieve pose estimation results for tracked frames.
        
        Returns camera position and orientation data.
      operationId: listPoses
      tags: [Poses]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of poses to return
        - name: frame_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific frame ID
      responses:
        '200':
          description: List of pose estimates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoseList'
  /poses/{pose_id}:
    get:
      summary: Get pose details
      description: Retrieve detailed information about a specific pose estimate
      operationId: getPose
      tags: [Poses]
      parameters:
        - name: pose_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier for the pose
      responses:
        '200':
          description: Pose details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pose'
...
components:
  schemas:
...
    PoseList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PoseSummary'
        total:
          type: integer
          example: 1500
    PoseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440000"
        frame_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        timestamp:
          type: number
          format: float
          example: 1695328058.123
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.95
...
    Pose:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/PoseSummary/id'
        frame_id:
          $ref: '#/components/schemas/PoseSummary/frame_id'
        timestamp:
          $ref: '#/components/schemas/PoseSummary/timestamp'
        position:
          type: object
          properties:
            x:
              type: number
              example: 12.345
            y:
              type: number
              example: -8.765
            z:
              type: number
              example: 45.678
        orientation:
          type: object
          properties:
            qx:
              type: number
              example: 0.999
            qy:
              type: number
              example: 0.012
            qz:
              type: number
              example: -0.045
            qw:
              type: number
              example: 0.987
        confidence:
          $ref: '#/components/schemas/PoseSummary/confidence'
...